refactor(core): fully integrate Clock trait in production code
The Clock abstraction was implemented but not integrated fully in all
parts of the code. Now we can fully enable deterministic testing with
full control over time-dependant logic.

- Move TestClock implementation from kapsel-testing to kapsel-core
- Replace all "real" time handling logic with the clock abstraction
- Remove any RealClock usage outside of main.rs (production path)
- TestEnv is now fully built around Clock abstraction
- Add AppState to API handlers to provide the clock abstraction for
timestamps.

Note: This commit is only concerned with integrating the correct
behavior, we have failing tests due to additional manual adjustments
needed in tests and known problems with the test harness lifeycle
handling which causes test data contamination, these will be adressed in
follow-up commits.