feat(testing): hybrid connection strategy
The current testing harness puts extreme pressure on the postgres
connection pool. This forces us to increase the max_connections to 1000,
which in turn exhausts other resources, forcing us to reduce the nextest
thread usage to 4.

The idea here is to prevent this all these problems by providing the
correct APIs for the correct usages:

- Pure Logic / HTTP Client: TestUtilities::new() (no database needed)
- Standard DB Integration: TestConnection::new() (one connection)
- Concurrent DB Integration: TestPool::new_small() (small, private pool)
- Full System/Worker Tests: TestEnv::run_isolated_test() (unchanged, need
  commited data)